requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
pydantic[email]
python-multipart

# Baza Danych (Async)
sqlalchemy
asyncpg
aiosqlite

# Bezpieczeństwo
passlib
bcrypt==3.2.0
python-jose[cryptography]

# Testowanie
pytest
pytest-asyncio
httpx

# Modele i Dane
apscheduler
pmdarima
arch
yfinance
